// ui/src/generated/SessionReply.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Diagnostic } from "./Diagnostic";
import type { JsonValue } from "./serde_json/JsonValue";

export type SessionReply = { id: string | null, ok: boolean, result?: JsonValue, error?: string, diagnostic?: Diagnostic, };