// ui/src/generated/CapabilityDescriptor.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CancellationClass } from "./CancellationClass";
import type { CapabilityDocumentation } from "./CapabilityDocumentation";
import type { CapabilityKind } from "./CapabilityKind";
import type { CapabilityRef } from "./CapabilityRef";
import type { EffectHint } from "./EffectHint";
import type { IdempotencyClass } from "./IdempotencyClass";
import type { RetryClass } from "./RetryClass";
import type { JsonValue } from "./serde_json/JsonValue";

export type CapabilityDescriptor = { kind: CapabilityKind, capability: CapabilityRef, domain: string, input_schema: JsonValue, output_schema: JsonValue, recovery_schema: JsonValue, documentation: CapabilityDocumentation, required_scopes: Array<string>, potential_effects: Array<EffectHint>, idempotency: IdempotencyClass, retry: RetryClass, cancellation: CancellationClass, };